refactor made to use mockwebserver
