--- app/src/androidTest/java/br/com/concrete/tentacle/ExampleInstrumentedTest.kt
@@ -1,21 +1,0 @@
-package br.com.concrete.tentacle.tentacle
-
-import androidx.test.InstrumentationRegistry
-import org.junit.Assert.assertEquals
-import org.junit.Ignore
-import org.junit.Test
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@Ignore
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("br.com.concrete.tentacle.tentacle", appContext.packageName)
-    }
-}
