+++ app/src/test/java/br/com/concrete/tentacle/base/BaseViewModelTest.kt
@@ -2,15 +2,18 @@ package br.com.concrete.tentacle.base
 
 import br.com.concrete.tentacle.di.PROPERTY_BASE_URL
 import br.com.concrete.tentacle.rules.RxImmediateSchedulerRule
+import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.After
 import org.junit.Before
+import org.junit.Ignore
 import org.junit.Rule
 import org.junit.runner.RunWith
 import org.koin.standalone.StandAloneContext
 import org.koin.test.KoinTest
 import org.robolectric.RobolectricTestRunner
 
+@Ignore
 @RunWith(RobolectricTestRunner::class)
 open class BaseViewModelTest : KoinTest {
 
@@ -29,5 +32,21 @@ open class BaseViewModelTest : KoinTest {
     @Throws fun tearDown() {
         StandAloneContext.stopKoin()
         mockServer.shutdown()
+    }
+
+    fun getJson(path: String): String? {
+        return this::class
+            .java
+            .classLoader?.getResource(
+            path
+        )?.readText()
+    }
+
+    fun mockResponseError400(responseJson: String?) {
+        val mockResponse = MockResponse()
+            .setResponseCode(400)
+            .setBody(responseJson)
+
+        mockServer.enqueue(mockResponse)
     }
 }
